fixed BIGINT for tg id for postgresql :/

+++ models/activities.py
@@ -1,6 +1,7 @@
 from typing import TYPE_CHECKING
 
-from sqlalchemy import ForeignKey, BigInteger
+from sqlalchemy import ForeignKey
+from sqlalchemy.dialects.postgresql import BIGINT
 from sqlalchemy.orm import Mapped, relationship, mapped_column
 
 if TYPE_CHECKING:
@@ -14,7 +15,7 @@ from dtos import ActivityDto
 class Activity(Base):
     __tablename__ = 'activities'
     user_id: Mapped[int] = mapped_column(
-        BigInteger,
+        BIGINT,
         ForeignKey('users.id', ondelete='CASCADE'),
         primary_key=True
     )
