+++ models/users.py
@@ -1,7 +1,8 @@
 from typing import TYPE_CHECKING
 
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from sqlalchemy import String, BigInteger
+from sqlalchemy.dialects.postgresql import BIGINT
+from sqlalchemy import String
 
 if TYPE_CHECKING:
     from .activities import Activity
@@ -11,7 +12,7 @@ from config import Base
 
 class User(Base):
     __tablename__ = 'users'
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
+    id: Mapped[int] = mapped_column(BIGINT, primary_key=True)
     email: Mapped[str] = mapped_column(String(320), unique=True)
     fullname: Mapped[str] = mapped_column(String(244))
